move EventHubClient into its own file
