--- node/tasks
@@ -1,1 +1,0 @@
-* move EventHubClient into its own file
